requirements.txt
certifi==2018.1.18
Django==2.0.2
django-cors-headers==2.1.0
djangorestframework==3.7.7
idna==2.6
pytz==2017.3
requests==2.18.4
django-rest-swagger==2.1.2
python-docx==0.8.6
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.0
django-reversion==2.0.13
python-crontab==2.2.8
PyMySQL==0.8.1
XlsxWriter==1.0.5
PyMySQL==0.8.1
python-docx==0.8.6
pypinyin==0.36.0
selenium==3.141.0
jsonpath==0.82
django-apscheduler==0.3.0
pymssql ==2.1.4
